Redis 관련 컨트롤러 분리
