+++ src/main/java/team/shdsesc/stocksimul/agent/PredictResponseDTO.java
@@ -23,4 +23,18 @@ public class PredictResponseDTO {
     @JsonProperty("investment_analysis")
     private InvestmentAnalysisDTO investmentAnalysis;
 
+    // Chat.jsx에서 시뮬레이션 결과에 사용됨
+    @JsonProperty("last_price")
+    private Double lastPrice;
+
+    // 차트용 예측 데이터 (Chat.jsx 모달에서 사용)
+    @JsonProperty("prediction_dates")
+    private java.util.List<String> predictionDates;
+
+    @JsonProperty("price_predictions")
+    private java.util.List<Double> pricePredictions;
+
+    @JsonProperty("return_predictions")
+    private java.util.List<Double> returnPredictions;
+
 }
